store/requirements.txt
Django>=5.2.1
django-environ>=0.11.2
gunicorn>=21.2.0
django-imagekit>=4.1.0 